Moved to JDK 1.8.0 and added some generally needed tools.
There is a push on the Jenkins core team to move to Java 8 .
They dropped the Java 6 support with 1.608 and there was a plan
to move straight to Java 8. While this is not a requirement now,
it seems to be a good idea to go to the latest stable version.